requirements.txt
python-docx==0.8.10
prettytable==0.7.2
PyPDF2==1.26.0
xlsxwriter==1.3.3
pdfminer.six